Move `MarkdownLine` to `Icinga\Module\Icingadb\Widget`
